projectlist prototype
